+++ cases/graph_patches_classification.py
@@ -4,14 +4,15 @@ import datetime
 from typing import List, Union
 
 import numpy as np
-from nas.var import PROJECT_ROOT, VERBOSE_VAL
-from nas.patches.utils import set_root, set_tf_compat
+from nas.utils.var import PROJECT_ROOT, VERBOSE_VAL
+from nas.utils.utils import set_root, set_tf_compat
 
 from fedot.core.repository.quality_metrics_repository import MetricsRepository, ClassificationMetricsEnum
 
-from nas.patches.load_images import from_images
-from nas.graph_cnn_gp_operators import generate_initial_graph
-from nas.composer.graph_gp_cnn_composer import GPNNGraphOptimiser, GPNNComposerRequirements
+from nas.utils.load_images import from_images
+from nas.composer.cnn_graph_operator import generate_initial_graph
+from nas.composer.gp_cnn_optimiser import GPNNGraphOptimiser
+from nas.composer.gp_cnn_composer import GPNNComposerRequirements
 
 from fedot.core.dag.validation_rules import has_no_cycle, has_no_self_cycled_nodes
 from fedot.core.log import default_log
@@ -26,7 +27,9 @@ from fedot.core.optimisers.gp_comp.operators.mutation import single_edge_mutatio
     single_change_mutation, single_drop_mutation
 from nas.composer.metrics import calculate_validation_metric
 
-from nas.composer.graph_gp_cnn_composer import NNGraph, CustomGraphAdapter, NNNode
+from nas.composer.cnn_adapters import CustomGraphAdapter
+from nas.composer.cnn_graph_node import NNNode
+from nas.composer.cnn_graph import NNGraph
 
 root = PROJECT_ROOT
 set_root(root)
