+++ tests/unit/nas/test_keras_eval.py
@@ -1,7 +1,9 @@
-from nas.composer.graph_gp_cnn_composer import NNGraph, NNNode, GPNNComposerRequirements
-from nas.graph_cnn_gp_operators import random_conv_graph_generation
-from nas.graph_keras_eval import create_nn_model
-from nas.patches.utils import project_root
+from nas.composer.gp_cnn_composer import GPNNComposerRequirements
+from nas.composer.cnn_graph_node import NNNode
+from nas.composer.cnn_graph import NNGraph
+from nas.composer.cnn_graph_operator import random_conv_graph_generation
+from nas.nn.graph_keras_eval import create_nn_model
+from nas.utils.utils import project_root
 
 root = project_root()
 requirements = GPNNComposerRequirements(image_size=[120, 120], init_graph_with_skip_connections=True)
