+++ tests/unit/nas/test_validation.py
@@ -2,9 +2,10 @@ import os
 
 from nas.graph_cnn_mutations import has_no_flatten_skip, graph_has_several_starts, flatten_check, \
     graph_has_wrong_structure
-from nas.composer.graph_gp_cnn_composer import NNGraph, NNNode
-from nas.graph_cnn_gp_operators import generate_initial_graph
-from nas.var import TESTING_ROOT
+from nas.composer.cnn_graph_node import NNNode
+from nas.composer.cnn_graph import NNGraph
+from nas.composer.cnn_graph_operator import generate_initial_graph
+from nas.utils.var import TESTING_ROOT
 
 
 def generate_graph():
